+++ neo/Utils/BlockchainFixtureTestCase.py
@@ -10,12 +10,15 @@ from neo.Settings import settings
 from neo.logging import log_manager
 
 logger = log_manager.getLogger()
+import neo
 
 
 class BlockchainFixtureTestCase(NeoTestCase):
     FIXTURE_REMOTE_LOC = 'https://s3.us-east-2.amazonaws.com/cityofzion/fixtures/fixtures_v8.tar.gz'
     FIXTURE_FILENAME = os.path.join(settings.DATA_DIR_PATH, 'Chains/fixtures_v8.tar.gz')
     _blockchain = None
+
+    root_folder = os.path.dirname(neo.__file__)[:-4]
 
     @classmethod
     def leveldb_testpath(cls):