+++ neo/bin/prompt.py
@@ -30,7 +30,7 @@ from neo.Prompt.Commands.BuildNRun import BuildAndRun, LoadAndRun
 from neo.Prompt.Commands.Invoke import InvokeContract, TestInvokeContract, test_invoke
 from neo.Prompt.Commands.LoadSmartContract import LoadContract, GatherContractDetails, ImportContractAddr, \
     ImportMultiSigContractAddr
-from neo.Prompt.Commands.Send import construct_and_send, parse_and_sign
+from neo.Prompt.Commands.Send import construct_send_basic, construct_send_many, process_transaction, parse_and_sign
 
 from neo.Prompt.Commands.Tokens import token_approve_allowance, token_get_allowance, token_send, token_send_from, \
     token_mint, token_crowdsale_register, token_history
@@ -144,6 +144,7 @@ class PromptInterface:
                 'wallet split {addr} {asset} {unspent index} {divide into number of vins}',
                 'wallet close',
                 'send {assetId or name} {address} {amount} (--from-addr={addr}) (--fee={priority_fee})',
+                'sendmany {number of outgoing tx} (--change-addr={addr}) (--from-addr={addr}) (--fee={priority_fee})',
                 'sign {transaction in JSON format}',
                 'testinvoke {contract hash} [{params} or --i] (--attach-neo={amount}, --attach-gas={amount}) (--from-addr={addr}) --no-parse-addr (parse address strings to script hash bytearray)',
                 'debugstorage {on/off/reset}'
@@ -188,7 +189,7 @@ class PromptInterface:
                                 'watch_addr', 'contract_addr', 'testinvoke', 'tkn_send',
                                 'tkn_mint', 'tkn_send_from', 'tkn_approve', 'tkn_allowance',
                                 'tkn_register', 'build', 'notifications', 'tkn_history',
-                                'sign', 'send', 'withdraw', 'nep2', 'multisig_addr', 'token',
+                                'sign', 'send', 'sendmany', 'withdraw', 'nep2', 'multisig_addr', 'token',
                                 'claim', 'migrate', 'rebuild', 'create_addr', 'delete_addr',
                                 'delete_token', 'alias', 'unspent', 'split', 'close',
                                 'withdraw_reqest', 'completed', 'cancel', 'cleanup',
@@ -548,11 +549,18 @@ class PromptInterface:
             print("Wallet: '{}' is an invalid parameter".format(item))
 
     def do_send(self, arguments):
-        construct_and_send(self, self.Wallet, arguments)
+        framework = construct_send_basic(self.Wallet, arguments)
+        if type(framework) is list:
+            process_transaction(self.Wallet, contract_tx=framework[0], scripthash_from=framework[1], fee=framework[2], owners=framework[3], user_tx_attributes=framework[4])
+
+    def do_send_many(self, arguments):
+        framework = construct_send_many(self.Wallet, arguments)
+        if type(framework) is list:
+            process_transaction(self.Wallet, contract_tx=framework[0], scripthash_from=framework[1], scripthash_change=framework[2], fee=framework[3], owners=framework[4], user_tx_attributes=framework[5])
 
     def do_sign(self, arguments):
         jsn = get_arg(arguments)
-        parse_and_sign(self, self.Wallet, jsn)
+        parse_and_sign(self.Wallet, jsn)
 
     def show_state(self):
         height = Blockchain.Default().Height
@@ -973,6 +981,8 @@ class PromptInterface:
                         self.show_wallet(arguments)
                     elif command == 'send':
                         self.do_send(arguments)
+                    elif command == 'sendmany':
+                        self.do_send_many(arguments)
                     elif command == 'sign':
                         self.do_sign(arguments)
                     elif command == 'block':